capsules/content-generator-mcp/pyproject.toml
[build-system]
requires = ["setuptools>=70.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "alquimia-tech-content-generator-mcp"
version = "0.1.0"
description = "An MCP server to generate technical documents based on Alquimia API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiosseclient>=0.1.8",
    "fastmcp>=2.10.6",
    "httpx>=0.28.1",
    "langchain-text-splitters>=0.3.9",
    "loguru>=0.7.3",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["alquimia*"]

[dependency-groups]
dev = [
    "pdoc3>=0.11.6",
    "ruff>=0.12.9",
]

